added DropTables

+++ DosjunEditor/Data/Enum/MonsterFlags.cs
@@ -3,7 +3,7 @@
 namespace DosjunEditor
 {
     [Flags]
-    public enum MonsterFlags : byte
+    public enum MonsterFlags : ushort
     {
         None = 0,
         HasSkills = 1,
